// src/app/pages/corralons/components/corralons-table/corralons.interface.ts
export interface CorralonsInterface {
   idcorralon?: number;
   fecha?: string;
   infraccionNumero?: number;
   corralonNombre?: string;
   motivo?: string;
   status?: string;
   permisotaxiasignado_idpermisotaxiasignado?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}